oldVersion/requirements.txt
opencv-python
customtkinter
pillow